% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammacurve.R
\name{gammacurve}
\alias{gammacurve}
\title{Estimate Gamma Distribution Parameters and Compute PDF or CDF}
\usage{
gammacurve(data, variable, group = NULL, func = 1)
}
\arguments{
\item{data}{A data frame containing the variable to be analyzed.}

\item{variable}{A string or unquoted column name specifying the numeric variable
for which the Gamma distribution parameters should be estimated.}

\item{group}{Optional. A character vector of column names used to group the data
before parameter estimation. If \code{NULL} (default), parameters are estimated
using all data.}

\item{func}{An integer indicating which function values to return:
\describe{
\item{1}{Return the PDF (default).}
\item{2}{Return the CDF.}
}}
}
\value{
A tibble containing the original data with additional columns:
\describe{
\item{alpha_hat}{Estimated shape parameter of the Gamma distribution.}
\item{theta_hat}{Estimated scale parameter of the Gamma distribution.}
\item{PDF}{Gamma probability density function values (if \code{func = 1}).}
\item{CDF}{Gamma cumulative distribution function values (if \code{func = 2}).}
}
}
\description{
Estimates the shape (\code{alpha}) and scale (\code{theta}) parameters of a Gamma
distribution from a numeric variable using the method of moments, and
computes either the probability density function (PDF) or cumulative
distribution function (CDF) values for each observation. Optionally supports
grouped estimation when grouping variables are provided.
}
\examples{
# to install the package
if(!require(remotes)) install.packages("remotes")
if(!requireNamespace("gammacurve", quietly = TRUE)) {
  remotes::install_github("agronomy4future/gammacurve", force= TRUE)
}
library(remotes)
library(gammacurve)

# to upload dataset for practice
if(!require(remotes)) install.packages("readr")
library (readr)
github="https://raw.githubusercontent.com/agronomy4future/raw_data_practice/refs/heads/main/sweet_potato_weight.csv"
df=data.frame(read_csv(url(github),show_col_types = FALSE))

print(head(df,3))
      Season Nitrogen Block   weight
1 2020_Season       N1     I 111.8380
2 2020_Season       N1     I   9.7788
3 2020_Season       N1     I  57.3240

# Example 1: Estimate Gamma parameters and compute PDF without grouping
output= gammacurve(df, variable="weight",
                  func=1) #func=1, PDF / func=2, CDF
print(head(output,3))

# Example 2: Estimate Gamma parameters and compute PDF with grouping
output= gammacurve(df, variable="weight",
                  group=c("Season", "Nitrogen"),
                  func=1) #func=1, PDF / func=2, CDF
print(head(output,3))

Github: https://github.com/agronomy4future/gammacurve
Website: https://agronomy4future.com
Contact: kimjk@agronomy4future

}
